fix: convert hashCode() Int to Long for fingerprint
- Fix type mismatch in checkCacheValidity()
- Fix type mismatch in generateFingerprint()
- hashCode() returns Int, convert to Long with .toLong()

+++ Animasu/src/main/kotlin/com/Animasu/SyncMonitor.kt
@@ -162,7 +162,7 @@ abstract class SyncMonitor {
             
             if (titles.isEmpty()) return false
             
-            val newFingerprint = titles.hashCode()
+            val newFingerprint = titles.hashCode().toLong()
             newFingerprint == currentFingerprint
         } catch (e: Exception) {
             Log.e(TAG, "checkCacheValidity failed: ${e.message}")
@@ -180,7 +180,7 @@ abstract class SyncMonitor {
         val titles = withTimeout(CHECK_TIMEOUT) { 
             fetchTitles(provider, url)
         }
-        return titles.hashCode()
+        return titles.hashCode().toLong()
     }
 
     /**
